+++ ios/Classes/JPushPlugin.m
@@ -3,12 +3,14 @@
 #import <UserNotifications/UserNotifications.h>
 #endif
 
-#import <JPush/JPUSHService.h>
+#import "JPUSHService.h"
+#import "JGInforCollectionAuth.h"
 
 #define JPLog(fmt, ...) NSLog((@"| JPUSH | Flutter | iOS | " fmt), ##__VA_ARGS__)
 
 @interface NSError (FlutterError)
 @property(readonly, nonatomic) FlutterError *flutterError;
+
 @end
 
 @implementation NSError (FlutterError)
@@ -21,7 +23,9 @@
 
 
 #if defined(__IPHONE_10_0) && __IPHONE_OS_VERSION_MAX_ALLOWED >= __IPHONE_10_0
-@interface JPushPlugin ()<JPUSHRegisterDelegate>
+@interface JPushPlugin ()<JPUSHRegisterDelegate,JPUSHInAppMessageDelegate>
+//在前台时是否展示通知
+@property(assign, nonatomic) BOOL unShow;
 @end
 #endif
 
@@ -120,6 +124,8 @@ static NSMutableArray<FlutterResult>* getRidResults;
         result([@"iOS " stringByAppendingString:[[UIDevice currentDevice] systemVersion]]);
     } else if([@"setup" isEqualToString:call.method]) {
         [self setup:call result: result];
+    } else if([@"setUnShowAtTheForeground" isEqualToString:call.method]) {
+        [self setUnShowAtTheForeground:call result: result];
     } else if([@"applyPushAuthority" isEqualToString:call.method]) {
         [self applyPushAuthority:call result:result];
     } else if([@"setTags" isEqualToString:call.method]) {
@@ -136,12 +142,14 @@ static NSMutableArray<FlutterResult>* getRidResults;
         [self setAlias:call result:result];
     } else if([@"deleteAlias" isEqualToString:call.method]) {
         [self deleteAlias:call result:result];
+    } else if([@"getAlias" isEqualToString:call.method]) {
+        [self getAlias:call result:result];
     } else if([@"setBadge" isEqualToString:call.method]) {
         [self setBadge:call result:result];
     } else if([@"stopPush" isEqualToString:call.method]) {
         [self stopPush:call result:result];
     } else if([@"resumePush" isEqualToString:call.method]) {
-        JPLog(@"ios platform not support resume push.");
+        [self resumePush:call result:result];
         //[self applyPushAuthority:call result:result];
     } else if([@"clearAllNotifications" isEqualToString:call.method]) {
         [self clearAllNotifications:call result:result];
@@ -157,6 +165,12 @@ static NSMutableArray<FlutterResult>* getRidResults;
         [self isNotificationEnabled:call result:result];
     } else if([@"openSettingsForNotification"isEqualToString:call.method]) {
         [self openSettingsForNotification];
+    } else if ([@"setAuth" isEqualToString:call.method]) {
+        [self setAuth:call result:result];
+    } else if ([@"pageEnterTo" isEqualToString:call.method]) {
+        [self pageEnterTo:call];
+    } else if ([@"pageLeave" isEqualToString:call.method]) {
+        [self pageLeave:call];
     } else{
         result(FlutterMethodNotImplemented);
     }
@@ -175,10 +189,20 @@ static NSMutableArray<FlutterResult>* getRidResults;
         [JPUSHService setLogOFF];
     }
     
+    [JPUSHService setInAppMessageDelegate:self];
+    
     [JPUSHService setupWithOption:_completeLaunchNotification
                            appKey:arguments[@"appKey"]
                           channel:arguments[@"channel"]
                  apsForProduction:[arguments[@"production"] boolValue]];
+}
+
+//设置APP在前台时是否展示通知
+- (void)setUnShowAtTheForeground:(FlutterMethodCall*)call result:(FlutterResult)result {
+    JPLog(@"setUnShowDidEnterBackground:");
+    NSDictionary *arguments = call.arguments;
+    NSNumber *unShow = arguments[@"UnShow"];
+    if(unShow && [unShow isKindOfClass:[NSNumber class]]) self.unShow = [unShow boolValue];
 }
 
 - (void)applyPushAuthority:(FlutterMethodCall*)call result:(FlutterResult)result {
@@ -302,6 +326,18 @@ static NSMutableArray<FlutterResult>* getRidResults;
     } seq: 0];
 }
 
+- (void)getAlias:(FlutterMethodCall*)call result:(FlutterResult)result {
+    JPLog(@"getAlias:%@",call.arguments);
+    [JPUSHService getAlias:^(NSInteger iResCode, NSString *iAlias, NSInteger seq) {
+        if (iResCode == 0) {
+            result(@{@"alias": iAlias ?: @""});
+        }else {
+            NSError *error = [[NSError alloc] initWithDomain:@"JPush.Flutter" code:iResCode userInfo:nil];
+            result([error flutterError]);
+        }
+    } seq: 0];
+}
+
 - (void)setBadge:(FlutterMethodCall*)call result:(FlutterResult)result {
     JPLog(@"setBadge:%@",call.arguments);
     NSInteger badge = [call.arguments[@"badge"] integerValue];
@@ -316,6 +352,12 @@ static NSMutableArray<FlutterResult>* getRidResults;
     JPLog(@"stopPush:");
     [[UIApplication sharedApplication] unregisterForRemoteNotifications];
 }
+
+- (void)resumePush:(FlutterMethodCall*)call result:(FlutterResult)result {
+    JPLog(@"resumePush:");
+    [[UIApplication sharedApplication] registerForRemoteNotifications];
+}
+
 - (void)clearAllNotifications:(FlutterMethodCall*)call result:(FlutterResult)result {
     JPLog(@"clearAllNotifications:");
     
@@ -404,8 +446,13 @@ static NSMutableArray<FlutterResult>* getRidResults;
         content.userInfo = params[@"extra"];
     }
     
-    if (params[@"sound"] && ![params[@"sound"] isEqualToString:@"<null>"]) {
-        content.sound = params[@"sound"];
+    if (params[@"soundName"] && ![params[@"soundName"] isEqualToString:@"<null>"]) {
+        content.sound = params[@"soundName"];
+    }
+    
+    if (@available(iOS 15.0, *)) {
+      content.interruptionLevel = UNNotificationInterruptionLevelActive;
+      content.relevanceScore = 1;
     }
     
     JPushNotificationTrigger *trigger = [[JPushNotificationTrigger alloc] init];
@@ -466,6 +513,26 @@ static NSMutableArray<FlutterResult>* getRidResults;
     }];
 }
 
+- (void)setAuth:(FlutterMethodCall*)call result:(FlutterResult)result {
+    JPLog(@"setAuth:%@",call.arguments);
+    BOOL enable = [call.arguments[@"enable"] boolValue];
+    [JGInforCollectionAuth JCollectionAuth:^(JGInforCollectionAuthItems * _Nonnull authInfo) {
+        authInfo.isAuth = enable;
+    }];
+    
+}
+
+- (void)pageEnterTo:(FlutterMethodCall*)call {
+    JPLog(@"pageEnterTo:%@",call.arguments);
+    NSString *pageName = call.arguments;
+    [JPUSHService pageEnterTo:pageName];
+}
+
+- (void)pageLeave:(FlutterMethodCall*)call {
+    JPLog(@"pageLeave:%@",call.arguments);
+    NSString *pageName = call.arguments;
+    [JPUSHService pageLeave:pageName];
+}
 
 
 - (void)dealloc {
@@ -559,9 +626,28 @@ static NSMutableArray<FlutterResult>* getRidResults;
 
 - (BOOL)application:(UIApplication *)application didReceiveRemoteNotification:(NSDictionary *)userInfo fetchCompletionHandler:(void (^)(UIBackgroundFetchResult))completionHandler {
     JPLog(@"application:didReceiveRemoteNotification:fetchCompletionHandler");
-    
     [JPUSHService handleRemoteNotification:userInfo];
-    [_channel invokeMethod:@"onReceiveNotification" arguments:userInfo];
+    if (@available(* ,iOS 10)) {
+        [_channel invokeMethod:@"onReceiveNotification" arguments:userInfo];
+       
+        /**
+         * 下面这段代码是解决 app处于杀死状态,点击通知启动app,但是不回调onOpenNotification的问题。
+         * 上诉情况不会走didReceiveNotificationResponse：回调。但是会走didReceiveRemoteNotification:fetchCompletionHandler:回调。iOS原生项目中正常情况下，点击通知冷启动app是会回调didReceiveNotificationResponse，不回调didReceiveRemoteNotification:fetchCompletionHandler:的。
+         * 因为不走didReceiveNotificationResponse：回调 所以没有onOpenNotification回调。这跟生命周期有关，didReceiveNotificationResponse:的代理需要通知的远程代理设置要在didFinishLaunch结束之前。但是flutter初始化jpush是在didFinishLaunch之后。
+         * 在这个方法里做一个判断吧，如果收到的消息和启动时的消息是同一个消息，则判断该消息为app杀死状态下通过点击通知唤醒的。
+         */
+        
+        if (_launchNotification && userInfo && [_launchNotification isKindOfClass:[NSDictionary class]] && [userInfo isKindOfClass:[NSDictionary class]]) {
+            // 拿到启动时的推送数据里的msgid
+            NSNumber *launchMsgid = [_launchNotification valueForKey:@"_j_msgid"];
+            // 拿到收到的消息的msgid
+            NSNumber *msgid = [userInfo valueForKey:@"_j_msgid"];
+            // 如果消息id一致
+            if ([launchMsgid isKindOfClass:[NSNumber class]] && [msgid isKindOfClass:[NSNumber class]] && [[launchMsgid stringValue] isEqualToString:[msgid stringValue]]) {
+                [_channel invokeMethod:@"onOpenNotification" arguments:_launchNotification];
+            }
+        }
+    }
     completionHandler(UIBackgroundFetchResultNewData);
     return YES;
 }
@@ -590,17 +676,19 @@ static NSMutableArray<FlutterResult>* getRidResults;
     });
 }
 
+//前台收到本地通知
 - (void)jpushNotificationCenter:(UNUserNotificationCenter *)center willPresentNotification:(UNNotification *)notification withCompletionHandler:(void (^)(NSInteger))completionHandler  API_AVAILABLE(ios(10.0)){
     JPLog(@"jpushNotificationCenter:willPresentNotification::");
     NSDictionary * userInfo = notification.request.content.userInfo;
     if([notification.request.trigger isKindOfClass:[UNPushNotificationTrigger class]]) {
         [JPUSHService handleRemoteNotification:userInfo];
-        [_channel invokeMethod:@"onReceiveNotification" arguments: [self jpushFormatAPNSDic:userInfo]];
+        if (@available(iOS 10 , *)) {
+            [_channel invokeMethod:@"onReceiveNotification" arguments: [self jpushFormatAPNSDic:userInfo]];
+        }
     }else{
         JPLog(@"iOS10 前台收到本地通知:userInfo：%@",userInfo);
     }
-    
-    completionHandler(notificationTypes);
+    if (!self.unShow) completionHandler(notificationTypes);
 }
 
 - (void)jpushNotificationCenter:(UNUserNotificationCenter *)center didReceiveNotificationResponse:(UNNotificationResponse *)response withCompletionHandler:(void (^)())completionHandler  API_AVAILABLE(ios(10.0)){
@@ -670,5 +758,28 @@ static NSMutableArray<FlutterResult>* getRidResults;
     formatDic[@"extras"] = extras;
     return formatDic;
 }
+
+
+#pragma mark - 应用内消息回调
+- (void)jPushInAppMessageDidShow:(JPushInAppMessage *)inAppMessage {
+    [_channel invokeMethod:@"onInAppMessageShow" arguments: [self convertInappMsg:inAppMessage]];
+}
+
+- (void)jPushInAppMessageDidClick:(JPushInAppMessage *)inAppMessage {
+    [_channel invokeMethod:@"onInAppMessageClick" arguments: [self convertInappMsg:inAppMessage]];
+}
+
+- (NSDictionary *)convertInappMsg:(JPushInAppMessage *)inAppMessage {
+    NSDictionary *result = @{
+        @"mesageId": inAppMessage.mesageId ?: @"",    // 消息id
+        @"title": inAppMessage.title ?:@"",       // 标题
+        @"content": inAppMessage.content ?: @"",    // 内容
+        @"target": inAppMessage.target ?: @[],      // 目标页面
+        @"clickAction": inAppMessage.clickAction ?: @"", // 跳转地址
+        @"extras": inAppMessage.extras ?: @{} // 附加字段
+    };
+    return result;
+}
+
 
 @end